Add new fn to use operate() & update old fns
Change operate to work with new symbols for multiply/divide. Rename changeDisplay() for clarification and have it use new fn checkButton() instead. Rename display() for clarification and parameter to be a string instead of node. Add 3 new global bool to help with storing values for checkButton() and change default values of other global var. checkButton() only updates display and stores values if button clicked in correct order (only 1 operator allowed, equal only works when there is 2 numbers and an operator)